+++ src/InternalException.ts
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: CC0-1.0
+
+import Exception from "./Exception";
+
+export default class InternalException extends Exception {
+	constructor(message: (string | null) = null,
+	            cause: (Exception | Error | null) = null) {
+		super(message, cause);
+	}
+}