+++ src/index.ts
@@ -2,6 +2,7 @@
 
 export { default, default as Exception } from "./Exception";
 export { default as IllegalStateException } from "./IllegalStateException";
+export { default as InternalException } from "./InternalException";
 export { default as InvalidArgumentException } from "./InvalidArgumentException";
 export { default as NoImplementationException } from "./NoImplementationException";
 export { default as OutOfRangeException } from "./OutOfRangeException";