+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [**Semantic Versioning v2.0.0**](https://semver.org/
 
 ## Unreleased ##
 
+### Added ###
+
+* pre-defined sub-exception: `InternalException`
+
 ### Security ###
 
 * Updated dependencies